GyreInABox

Documentation for GyreInABox.

Usage example

The below example illustrates setting up and running a simulation of the model with the default parameter values, and default configuration modulo reduction of the spatial resolution to $60\times 60 \times 25$, the simulation time to 6 hours and changing the model to record only a depth slice along longitude axis as outputs. The resulting depth slices of simulated fields are recorded as an animation using CairoMakie. The simulation time here is too short for the simulated fields to show any interesting behaviour, with this example just meant to illustrate the overall workflow while remaining relatively cheap to run.

using Oceananigans.Units
using GyreInABox

parameters = GyreInABoxParameters()
output_type = LongitudeDepthSlice()
configuration = GyreInABoxConfiguration(
    grid_size=(60, 60, 25),
    simulation_time=6hour,
    output_types=(output_type,)
)
run_simulation(parameters, configuration)
record_animation(configuration, output_type)

[ Info: Initializing simulation...
Iteration: 0000, time: 0 seconds, Δt: 15 minutes, wall time: 0 seconds
  max(|u|) = 3.1e-05 ms⁻¹, max(|v|) = 3.4e-05 ms⁻¹, max(|w|) = 2.6e-07 ms⁻¹, max(|S|) = 3.500e+01 g/kg, max(|T|) = 2.736e+01 ᵒC, max(|e|) = 0.000e+00 m²s⁻²
[ Info:     ... simulation initialization complete (44.824 seconds)
[ Info: Executing initial time step...
[ Info:     ... initial time step complete (27.079 seconds).
[ Info: Simulation is stopping after running for 1.256 minutes.
[ Info: Simulation time 6 hours equals or exceeds stop time 6 hours.
[ Info: Recording an animation of longitude_depth_slice to gyre_model_longitude_depth_slice.mp4...

Example simulated output

API reference

GyreInABox.GyreInABox

GyreInABox.DepthTimeAveraged

GyreInABox.GyreInABoxConfiguration

GyreInABox.GyreInABoxParameters

GyreInABox.HorizontalSlice

GyreInABox.LatitudeDepthSlice

GyreInABox.LongitudeDepthSlice

GyreInABox.add_output_writers!

GyreInABox.add_progress_message_callback!

GyreInABox.axis_aspect_ratio

GyreInABox.axis_limits

GyreInABox.axis_xlabel

GyreInABox.axis_ylabel

GyreInABox.bottom_meridional_drag

GyreInABox.bottom_zonal_drag

GyreInABox.boundary_conditions

GyreInABox.hyperbolically_spaced_faces

GyreInABox.indices

GyreInABox.initialize_model!

GyreInABox.label

GyreInABox.max_surface_wind_stress

GyreInABox.output_filename

GyreInABox.outputs

GyreInABox.record_animation

GyreInABox.reference_surface_temperature

GyreInABox.run_simulation

GyreInABox.schedule

GyreInABox.setup_grid

GyreInABox.setup_model

GyreInABox.setup_simulation

GyreInABox.surface_salinity_flux

GyreInABox.surface_temperature_flux

GyreInABox.thermocline_temperature_profile

GyreInABox.zonal_wind_stress

GyreInABox.GyreInABox — Module

Oceananigans based model of an ocean gyre in a bounded domain.

Details

Model of wind and buoyancy forced ocean gyre adapted from MITgcm baroclinic ocean gyre example from documentation implemented using Oceananigans (Ramadhan et al. 2020).

Compared to MITgcm example, beyond the change in the underlying software framework, some key differences in the (default) model configuration are

The model uses a seawater buoyancy formulation with tracer fields corresponding to salinity and (potential) temperature used to determine the buoyancy using a non-linear (polynomial) equation of state (Roquet et al. 2015).

A weighted essentially non-oscillatory numerical scheme is used for momentum advection (Silvestri et al. 2024).

The turbulence closure / parameterization used for small-scale turbulent mixing is a convective adjustment turbulent kinetic energy (CATKE) scheme (Wagner et al. 2025).

Adaptive time stepping is used based on controlling advective CFL number.

DepthTimeAveraged

GyreInABoxConfiguration

GyreInABoxParameters

HorizontalSlice

LatitudeDepthSlice

LongitudeDepthSlice

initialize_model!

record_animation

run_simulation

setup_model

setup_simulation

References

MITgcm contributors (2025). MITgcm/MITgcm: checkpoint69e (Version checkpoint69e). Zenodo. https://doi.org/10.5281/zenodo.15320163

Ramadhan, A., Wagner, G., Hill, C., Campin, J. M., Churavy, V., Besard, T., Souza, A., Edelman, A., Ferrari, R. & Marshall, J. (2020). Oceananigans. jl: Fast and friendly geophysical fluid dynamics on GPUs. Journal of Open Source Software, 5(53).

Roquet, F.; Madec, G.; Brodeau, L. and Nycander, J. (2015). Defining a simplified yet “realistic” equation of state for seawater. Journal of Physical Oceanography 45, 2564-2579.

Silvestri, S., Wagner, G. L., Campin, J. M., Constantinou, N. C., Hill, C. N., Souza, A., & Ferrari, R. (2024). A new WENO-based momentum advection scheme for simulations of ocean mesoscale turbulence. Journal of Advances in Modeling Earth Systems, 16(7), e2023MS004130.

Wagner, G. L., Hillier, A., Constantinou, N. C., Silvestri, S., Souza, A., Burns, K. J., Hill, C., Campin, J-M., Marshall, J. & Ferrari, R. (2025). Formulation and calibration of CATKE, a one-equation parameterization for microscale ocean mixing. Journal of Advances in Modeling Earth Systems, 17(4), e2024MS004522.

GyreInABox.DepthTimeAveraged — Type

struct DepthTimeAveraged{T} <: GyreInABox.AbstractHorizontalOutputType{T}

Depth and time averaged output.

Details

Records horizontal fields corresponding to depth averaged model velocity and traced fields temporally averaged over a specified window and a given temporal interval.

interval::Any: Time interval to record output at / s

average_window::Any: Time window to accumulate output averages over / s

GyreInABox.GyreInABoxConfiguration — Type

struct GyreInABoxConfiguration{T, A<:Oceananigans.Architectures.AbstractArchitecture, EOS, MA<:Oceananigans.Advection.AbstractAdvectionScheme, C<:Oceananigans.Coriolis.AbstractRotation, TC<:Oceananigans.TurbulenceClosures.AbstractTurbulenceClosure}

Configuration for ocean gyre model.

Details

Variables defining overall configuration of model such as spatial and temporal domains and corresponding discretizations, numerical schemes to use for different model components and outputs to record.

Compared to the GyreInABoxParameters the fields here mostly have discrete values and would not be typically calibrated to data.

architecture::Oceananigans.Architectures.AbstractArchitecture: Computational architecture to run simulation on

momentum_advection::Oceananigans.Advection.AbstractAdvectionScheme: Momentum advection scheme

equation_of_state::Any: Buoyancy equation of state

coriolis::Oceananigans.Coriolis.AbstractRotation: Coriolis force

closure::Oceananigans.TurbulenceClosures.AbstractTurbulenceClosure: Turbulence closure

tracers::Tuple: Tracer variables - needs to match turbulence closure and buoyancy formulation.

grid_size::Tuple: Grid dimensions in longitude, latitude and depth

longitude_interval::Tuple: Extent of spatial domain in longitude / °

latitude_interval::Tuple: Extent of spatial domain in latitude / °

depth_interval::Tuple: Extent of spatial domain in depth / m

depth_stretching_factor::Any: Stretching factor for hyperbolic spaced depth grid

halo_size::Tuple: Dimensions of grid halo region in longitude, latitude and depth

simulation_time::Any: Time to simulate for / s

initial_timestep::Any: Initial time step to use / s

maximum_timestep::Any: Maximum time step to use in time step adaptation / s

output_filename::String: Stem of output file names

progress_message_interval::Int64: Iteration interval between progress messages

random_seed::Int64: Random seed for state initialization

target_cfl::Any: Target (advective) CFL number for time stepping wizard

wizard_update_interval::Int64: Update (iteration) interval for time stepping wizard

wizard_max_change::Any: Maximum relative time step change in each wizard update

output_types::Tuple: Output types to record during simulation

GyreInABox.GyreInABoxParameters — Type

struct GyreInABoxParameters{T}

Parameters for ocean gyre model.

Details

Real-valued parameters of model controlling initial and boundary conditions.

u₁₀::Any: Average wind velocity 10 meters above the ocean / m s⁻¹

cᴰ::Any: Dimensionless drag coefficient

ρₐ::Any: Approximate average density of air at sea-level, kg m⁻³

φ₀::Any: Latitude offset for wind stress variation / °

Lφ::Any: Latitude scale for wind stress variation / °

μ::Any: Bottom drag damping / m s⁻¹

ρₒ::Any: Sea water density / kg m⁻³

Q::Any: Surface heat flux / W m⁻²

cᴾ::Any: Sea water specific heat capacity / J K⁻¹ kg⁻¹

h::Any: Ocean-atmosphere heat transfer coefficient / W m⁻² K⁻¹

T_pole::Any: Reference polar surface temperature / °C

T_equator::Any: Reference equatorial surface temperature / °C

T_deep::Any: Reference deep ocean temperature / °C

z₀::Any: Thermocline reference depth / m

ℓ::Any: Thermocline reference length scale / m

evaporation_rate::Any: Salinity surface evaporation rate / m s⁻¹

initial_salinity::Any: Initial constant salinity level / g kg⁻¹

initial_temperature_noise_scale::Any: Coefficient scaling amplitude of noise in initial temperature field / K

initial_velocity_noise_scale::Any: Coefficient scaling amplitude of noise in initial velocity field / m s⁻¹

GyreInABox.HorizontalSlice — Type

struct HorizontalSlice{T} <: GyreInABox.AbstractHorizontalOutputType{T}

Horizontal (latitude-longitude) slice output.

Details

Records horizontal slices through model velocity and tracer fields at specified depth and interval.

depth::Any: Depth of slice / m

interval::Any: Time interval to record output at / s

GyreInABox.LatitudeDepthSlice — Type

struct LatitudeDepthSlice{T} <: GyreInABox.AbstractVerticalOutputType{T}

Vertical (latitude-depth) slice output.

Details

Records vertical slices through model velocity and tracer fields at specified longitude and interval.

longitude::Any: Longitude of slice / °

interval::Any: Time interval to record output at / s

GyreInABox.LongitudeDepthSlice — Type

struct LongitudeDepthSlice{T} <: GyreInABox.AbstractVerticalOutputType{T}

Vertical (longitude-depth) slice output.

Details

Records vertical slices through model velocity and tracer fields at specified latitude and interval.

latitude::Any: Latitude of slice / °

interval::Any: Time interval to record output at / s

GyreInABox.add_output_writers! — Method

Register output writers for output types specified in configuration in simulation.

add_output_writers!(simulation, configuration)

GyreInABox.add_progress_message_callback! — Method

Add callback for progress updates as configured in configuration in simulation.

add_progress_message_callback!(simulation, configuration)

GyreInABox.axis_aspect_ratio — Method

Aspect ratio for field heatmaps.

axis_aspect_ratio(
    configuration,
    output_type::GyreInABox.AbstractOutputType
) -> Makie.AxisAspect

GyreInABox.axis_limits — Method

Axis limits for field heatmaps.

axis_limits(
    configuration::GyreInABoxConfiguration,
    _::GyreInABox.AbstractOutputType
) -> Tuple{Tuple, Tuple}

GyreInABox.axis_xlabel — Method

Label for horizontal axis for field heatmaps.

axis_xlabel(_::GyreInABox.AbstractOutputType) -> String

GyreInABox.axis_ylabel — Method

Label for vertical axis for field heatmaps.

axis_ylabel(_::GyreInABox.AbstractOutputType) -> String

GyreInABox.bottom_meridional_drag — Method

Bottom surface drag on meridional velocity component in m² s⁻².

bottom_meridional_drag(i, j, grid, clock, model_fields, p)

Details

Computes bottom surface drag at horizontal grid indices i and j for grid grid and model clock clock, with current model fields model_fields and parameters p.

GyreInABox.bottom_zonal_drag — Method

Bottom surface drag on zonal velocity component in m² s⁻².

bottom_zonal_drag(i, j, grid, clock, model_fields, p)

Details

Computes bottom surface drag at horizontal grid indices i and j for grid grid and model clock clock, with current model fields model_fields and parameters p.

GyreInABox.boundary_conditions — Method

Construct named tuple of boundary conditions for model given parameters parameters.

boundary_conditions(parameters)

GyreInABox.hyperbolically_spaced_faces — Method

Compute hyperbolically spaced depth grid face coordinate at index k given model configuration configuration.

hyperbolically_spaced_faces(k, configuration)

GyreInABox.indices — Method

Spatial grid indices output type records fields at.

indices(
    type::GyreInABox.AbstractOutputType,
    grid
) -> Tuple{Colon, Colon, Any}

GyreInABox.initialize_model! — Method

Initialize state of ocean gyre model model with parameters parameters and random variables generated using random number generator rng.

initialize_model!(model, parameters, rng)

GyreInABox.label — Method

Symbol label for output type to use in naming output file and registering output writer.

label(_::GyreInABox.AbstractOutputType) -> Symbol

GyreInABox.max_surface_wind_stress — Method

Maximal surface wind stress given parameters p in m² s⁻².

max_surface_wind_stress(p)

GyreInABox.output_filename — Method

Filename to record outputs to.

output_filename(
    configuration::GyreInABoxConfiguration,
    label::Symbol,
    extension::String
) -> String

Details

Output filename stem is specified in configuration. For an output type type a label computed using label function is appended on to stem and file extension is specified by extension added.

GyreInABox.outputs — Method

Named tuple of output variables (fields) to record for output type.

outputs(_::GyreInABox.AbstractOutputType, model) -> Any

GyreInABox.record_animation — Method

Record an animation of fields recorded as model output.

record_animation(
    configuration,
    output_type;
    figure_size,
    frame_rate,
    u_limits,
    v_limits,
    w_limits,
    T_limits,
    S_limits,
    e_limits
)

Details

Generates and record to file an animation of model outputs for a model configuration configuration and output type output_type. The simulation must have already been run for this configuration and with specified output type active.

Animation is generated using CairoMakie with a figure window size in pixels of (figure_size), frame rate in frames per second of frame_rate and limits for the zonal velocity, meridional velocity, vertical velocity, temperature, salinity and turbulent kinetic energy field value color mappings specified by u_limits, v_limits, w_limits, T_limits, S_limits and e_limits respectively.

GyreInABox.reference_surface_temperature — Method

Reference surface temperature for restoring surface temperature boundary condition as function of latitude φand parametersp`.

reference_surface_temperature(φ, p)

GyreInABox.run_simulation — Method

Setup and initialize model then setup and run simulation with parameters parameters and configuration configuration.

run_simulation(parameters, configuration)

GyreInABox.schedule — Method

Time schedule to record output type at.

schedule(
    _::GyreInABox.AbstractOutputType
) -> Union{Oceananigans.Utils.TimeInterval{_A, Dates.DateTime} where _A, Oceananigans.Utils.TimeInterval{_A, <:Number} where _A}

GyreInABox.setup_grid — Method

Set up ocean gyre model grid with configuration configuration optionally overriding architecture in configuration with architecture.

setup_grid(configuration; architecture)

GyreInABox.setup_model — Method

Set up ocean gyre model with parameters parameters and configuration configuration.

setup_model(parameters, configuration)

GyreInABox.setup_simulation — Method

Set up simulation for model model with configuration configuration.

setup_simulation(model, configuration)

GyreInABox.surface_salinity_flux — Method

Top surface salinity evaporation rate in g kg⁻¹ m s⁻¹.

surface_salinity_flux(i, j, grid, clock, model_fields, p)

Details

Computes surface salinity evaporation rate at horizontal grid indices i and j for grid grid and model clock clock, with current model fields model_fields and parameters p.

GyreInABox.surface_temperature_flux — Method

Surface temperature flux given parameters p in K m s⁻¹.

surface_temperature_flux(i, j, grid, clock, model_fields, p)

Details

Relaxation boundary condition that restores surface temperature to latitude dependent profile determined by reference_surface_temperature with relaxation rate h / (ρ₀ * cₚ) where h is the heat transfer coefficient, ρ₀ seawater density and cₚ seawater specific heat capacity.

GyreInABox.thermocline_temperature_profile — Method

Thermocline temperature profile at latitude φ and depth z with parameters p.

thermocline_temperature_profile(φ, z, p)

Details

Computes a latitude and depth dependent temperature with thermocline like profile which smoothly varies from a latitude dependent surface temperature to a constant deep ocean temperature with a sigmoidal profile.

GyreInABox.zonal_wind_stress — Method

Zonal wind stress applied as surface (flux) boundary condition to velocity field.

zonal_wind_stress(λ, φ, t, p)

Details

Computes wind stress in zonal direction at longitude λ and latitude φ (both in degrees), time t and with model parameters p.